mod refactor
